Page Separation and Minor fixes
